Temperature workflow complete

+++ src/main/java/calculator/LocalProduceCalculator.java
@@ -8,7 +8,7 @@ public class LocalProduceCalculator {
      * @param weight Amount of gram food bought.
      * @return (Amount of grams bought imported food) - (Amount of grams bought local produced food)
      */
-    public static double produce(int weight) {
+    public static Double produce(int weight) {
         //Calculate amount of kg CO2 saved.
         Double localFood = 232.0 * weight;
         Double importFood = 560.0 * weight;
@@ -17,7 +17,7 @@ public class LocalProduceCalculator {
 
         doubleResult = doubleResult / 1000000;
         doubleResult = Math.round(doubleResult * 100.00) / 100.00;
-        return doubleResult.intValue();
+        return doubleResult;
     }
